pyproject.toml
[project]
name = "gamify"
version = "0.1.0"
description = "Pygame abstraction framework for rapid game development"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gamify"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
